Implementation attempt of Presenter (Supervising Controller) pattern. Added RetroLambda library and updated viewadapter to use lambda expressions; cleaner code!

+++ app/src/main/java/achan/nl/uitstelgedrag/persistence/datasources/TaskDataSource.java
@@ -101,6 +101,10 @@ public class TaskDataSource implements DataSource<Task> {
     @Override
     public boolean delete(int id) {
         Task task = get(id);
+
+        if (task == null)
+            return false;
+
         String query = "DELETE FROM Tasks WHERE id = " + task.id + "";
         database.execSQL(query);
         Log.i("UITSTELGEDRAG", "Deleted task " + task.id + ":" + task.description);
